API

API

Mongoc.BSON — Type.

BSON is a wrapper for C struct bson_t.

source

Mongoc.BSONError — Type.

Mirrors C struct bson_error_t.

BSONError instances addresses are passed to libmongoc API using Ref{BSONError}(error), and are owned by the Julia process.

typedef struct {
   uint32_t domain;
   uint32_t code;
   char message[504];
} bson_error_t;

source

Mongoc.Client — Type.

Client is a wrapper for C struct mongoc_client_t.

source

Mongoc.Collection — Type.

Collection is a wrapper for C struct mongoc_collection_t.

source

Mongoc.Cursor — Type.

Cursor is a wrapper for C struct mongoc_cursor_t.

source

Mongoc.Database — Type.

Database is a wrapper for C struct mongoc_database_t.

source

Mongoc.URI — Type.

URI is a wrapper for C struct mongoc_uri_t.

source

Mongoc.as_json_string — Method.

as_json_string(bson::BSON; canonical::Bool=false) :: String

Converts a bson object to a JSON string.

Example

julia> document = Mongoc.BSON("{ "hey" : 1 }")
Mongoc.BSON(Ptr{Nothing} @0x00007fbc8e62cc30)

julia> Mongoc.as_json_string(document)
"{ "hey" : 1 }"

julia> Mongoc.as_json_string(document, canonical=true)
"{ "hey" : { "$numberInt" : "1" } }"

C API

bson_as_canonical_extended_json

`bsonasrelaxedextendedjson

source

Mongoc.command_simple — Method.

command_simple(client::Client, database::String, command::Union{String, BSON}) :: BSON

Executes a command given by a JSON string or a BSON instance.

Example

julia> client = Mongoc.Client() # connects to localhost at port 27017
Client(URI("mongodb://localhost:27017"))

julia> bson_result = Mongoc.command_simple(client, "admin", "{ "ping" : 1 }")
Mongoc.BSON(Ptr{Nothing} @0x00007f8663e0d8d0)

julia> println(Mongoc.as_json_string(bson_result))
{ "ok" : 1.0 }

See also: command_simple_as_json.

source

Mongoc.command_simple_as_json — Method.

command_simple_as_json(client::Client, database::String, command::Union{String, BSON}) :: String

Same as command_simple, but returns a JSON string.

Example

julia> client = Mongoc.Client() # connects to localhost at port 27017
Client(URI("mongodb://localhost:27017"))

julia> result = Mongoc.command_simple_as_json(client, "admin", "{ "ping" : 1 }")
"{ "ok" : 1.0 }"

source

Mongoc.set_appname! — Method.

set_appname!(client::Client, appname::String)

Sets the application name for this client.

This string, along with other internal driver details, is sent to the server as part of the initial connection handshake.

C API

mongoc_client_set_appname.

source