API Reference

API Reference

Mongoc.QUERY_FLAG_AWAIT_DATA — Constant.

Use with MONGOCQUERYTAILABLE_CURSOR. Block rather than returning no data. After a period, time out.

source

Mongoc.QUERY_FLAG_EXHAUST — Constant.

Stream the data down full blast in multiple “reply” packets. Faster when you are pulling down a lot of data and you know you want to retrieve it all.

source

Mongoc.QUERY_FLAG_NONE — Constant.

Specify no query flags.

source

Mongoc.QUERY_FLAG_NO_CURSOR_TIMEOUT — Constant.

The server normally times out an idle cursor after an inactivity period (10 minutes). This prevents that.

source

Mongoc.QUERY_FLAG_OPLOG_REPLAY — Constant.

Used internally by MongoDB.

source

Mongoc.QUERY_FLAG_PARTIAL — Constant.

Get partial results from mongos if some shards are down (instead of throwing an error).

source

Mongoc.QUERY_FLAG_SLAVE_OK — Constant.

Allow query of replica set secondaries.

source

Mongoc.QUERY_FLAG_TAILABLE_CURSOR — Constant.

Cursor will not be closed when the last data is retrieved. You can resume this cursor later.

source

Mongoc.BSON — Type.

BSON is a wrapper for C struct bson_t.

source

Mongoc.BSONCode — Type.

BSON element with JavaScript source code.

source

Mongoc.BSONError — Type.

Mirrors C struct bson_error_t and can be allocated in the stack.

BSONError instances addresses are passed to libbson/libmongoc API using Ref(error), and are owned by the Julia process.

typedef struct {
   uint32_t domain;
   uint32_t code;
   char message[504];
} bson_error_t;

source

Mongoc.BSONIter — Type.

BSONIter mirrors C struct bsonitert and can be allocated in the stack.

According to libbson documentation, it is meant to be used on the stack and can be discarded at any time as it contains no external allocation. The contents of the structure should be considered private and may change between releases, however the structure size will not change.

Inspecting its size in C, we get:

sizeof(bson_iter_t) == 80

source

Mongoc.BSONObjectId — Type.

Mirrors C struct bson_oid_t.

BSONObjectId instances addresses are passed to libbson/libmongoc API using Ref(oid), and are owned by the Julia process.

typedef struct {
   uint8_t bytes[12];
} bson_oid_t;

source

Mongoc.BSONType — Type.

BSONType mirrors C enum bsontypet.

source

Mongoc.Client — Type.

Client is a wrapper for C struct mongoc_client_t.

source

Mongoc.Collection — Type.

Collection is a wrapper for C struct mongoc_collection_t.

source

Mongoc.Cursor — Type.

Cursor is a wrapper for C struct mongoc_cursor_t.

source

Mongoc.Database — Type.

Database is a wrapper for C struct mongoc_database_t.

source

Mongoc.QueryFlags — Type.

Mirrors C struct mongoc_query_flags_t.

These flags correspond to the MongoDB wire protocol. They may be bitwise or’d together. They may modify how a query is performed in the MongoDB server.

From: http://mongoc.org/libmongoc/current/mongocqueryflags_t.html

source

Mongoc.URI — Type.

URI is a wrapper for C struct mongoc_uri_t.

source

Mongoc.as_json — Method.

as_json(bson::BSON; canonical::Bool=false) :: String

Converts a bson object to a JSON string.

Example

julia> document = Mongoc.BSON("{ "hey" : 1 }")
BSON("{ "hey" : 1 }")

julia> Mongoc.as_json(document)
"{ "hey" : 1 }"

julia> Mongoc.as_json(document, canonical=true)
"{ "hey" : { "$numberInt" : "1" } }"

C API

bson_as_canonical_extended_json

bson_as_relaxed_extended_json

source

Mongoc.command_simple — Method.

command_simple(database::Database, command::Union{String, BSON}) :: BSON

Executes a command given by a JSON string or a BSON instance.

It returns the first document from the result cursor.

Example

julia> client = Mongoc.Client() # connects to localhost at port 27017
Client(URI("mongodb://localhost:27017"))

julia> bson_result = Mongoc.command_simple(client["admin"], "{ "ping" : 1 }")
BSON("{ "ok" : 1.0 }")

C API

mongoc_database_command_simple

source

Mongoc.find_one — Function.

find_one(collection::Collection, bson_filter::BSON=BSON(); options::Union{Nothing, BSON}=nothing) :: Union{Nothing, BSON}

Execute a query to a collection and returns the first element of the result set.

Returns nothing if the result set is empty.

source

Mongoc.set_appname! — Method.

set_appname!(client::Client, appname::String)

Sets the application name for this client.

This string, along with other internal driver details, is sent to the server as part of the initial connection handshake.

C API

mongoc_client_set_appname.

source