API Reference

API Reference

Mongoc.QUERY_FLAG_AWAIT_DATA — Constant.

Use with MONGOCQUERYTAILABLE_CURSOR. Block rather than returning no data. After a period, time out.

source

Mongoc.QUERY_FLAG_EXHAUST — Constant.

Stream the data down full blast in multiple “reply” packets. Faster when you are pulling down a lot of data and you know you want to retrieve it all.

source

Mongoc.QUERY_FLAG_NONE — Constant.

Specify no query flags.

source

Mongoc.QUERY_FLAG_NO_CURSOR_TIMEOUT — Constant.

The server normally times out an idle cursor after an inactivity period (10 minutes). This prevents that.

source

Mongoc.QUERY_FLAG_OPLOG_REPLAY — Constant.

Used internally by MongoDB.

source

Mongoc.QUERY_FLAG_PARTIAL — Constant.

Get partial results from mongos if some shards are down (instead of throwing an error).

source

Mongoc.QUERY_FLAG_SLAVE_OK — Constant.

Allow query of replica set secondaries.

source

Mongoc.QUERY_FLAG_TAILABLE_CURSOR — Constant.

Cursor will not be closed when the last data is retrieved. You can resume this cursor later.

source

Mongoc.BSON — Type.

BSON is a wrapper for C struct bson_t.

source

Mongoc.BSONCode — Type.

BSON element with JavaScript source code.

source

Mongoc.BSONError — Type.

Mirrors C struct bson_error_t and can be allocated in the stack.

BSONError instances addresses are passed to libbson/libmongoc API using Ref(error), and are owned by the Julia process.

typedef struct {
   uint32_t domain;
   uint32_t code;
   char message[504];
} bson_error_t;

source

Mongoc.BSONIter — Type.

BSONIter mirrors C struct bsonitert and can be allocated in the stack.

According to libbson documentation, it is meant to be used on the stack and can be discarded at any time as it contains no external allocation. The contents of the structure should be considered private and may change between releases, however the structure size will not change.

Inspecting its size in C, we get:

sizeof(bson_iter_t) == 80

source

Mongoc.BSONObjectId — Type.

Mirrors C struct bson_oid_t.

BSONObjectId instances addresses are passed to libbson/libmongoc API using Ref(oid), and are owned by the Julia process.

typedef struct {
   uint8_t bytes[12];
} bson_oid_t;

source

Mongoc.BSONSubType — Type.

BSONSubType mirrors C enum bsonsubtypet.

source

Mongoc.BSONType — Type.

BSONType mirrors C enum bsontypet.

source

Mongoc.Client — Type.

Client is a wrapper for C struct mongoc_client_t.

source

Mongoc.Collection — Type.

Collection is a wrapper for C struct mongoc_collection_t.

source

Mongoc.Cursor — Type.

Cursor is a wrapper for C struct mongoc_cursor_t.

source

Mongoc.Database — Type.

Database is a wrapper for C struct mongoc_database_t.

source

Mongoc.QueryFlags — Type.

Mirrors C struct mongoc_query_flags_t.

These flags correspond to the MongoDB wire protocol. They may be bitwise or’d together. They may modify how a query is performed in the MongoDB server.

From: http://mongoc.org/libmongoc/current/mongocqueryflags_t.html

source

Mongoc.URI — Type.

URI is a wrapper for C struct mongoc_uri_t.

source

Mongoc.add_user — Function.

add_user(database::Database, username::String, password::String, roles::Union{Nothing, BSON}, custom_data::Union{Nothing, BSON}=nothing)

This function shall create a new user with access to database.

Warning: Do not call this function without TLS.

source

Mongoc.as_json — Method.

as_json(bson::BSON; canonical::Bool=false) :: String

Converts a bson object to a JSON string.

Example

julia> document = Mongoc.BSON("{ "hey" : 1 }")
BSON("{ "hey" : 1 }")

julia> Mongoc.as_json(document)
"{ "hey" : 1 }"

julia> Mongoc.as_json(document, canonical=true)
"{ "hey" : { "$numberInt" : "1" } }"

C API

bson_as_canonical_extended_json

bson_as_relaxed_extended_json

source

Mongoc.command_simple — Method.

command_simple(database::Database, command::Union{String, BSON}) :: BSON

Executes a command given by a JSON string or a BSON instance.

It returns the first document from the result cursor.

Example

julia> client = Mongoc.Client() # connects to localhost at port 27017
Client(URI("mongodb://localhost:27017"))

julia> bson_result = Mongoc.command_simple(client["admin"], "{ "ping" : 1 }")
BSON("{ "ok" : 1.0 }")

C API

mongoc_database_command_simple

source

Mongoc.find_one — Function.

find_one(collection::Collection, bson_filter::BSON=BSON(); options::Union{Nothing, BSON}=nothing) :: Union{Nothing, BSON}

Execute a query to a collection and returns the first element of the result set.

Returns nothing if the result set is empty.

source

Mongoc.get_server_mongodb_version — Method.

Queries the version for the MongoDB server instance.

source

Mongoc.has_user — Method.

has_user(database::Database, user_name::String) :: Bool

Checks if database has a user named user_name.

source

Mongoc.read_bson — Method.

read_bson(filepath::AbstractString) :: Vector{BSON}

Reads all BSON documents from a file located at filepath.

This will open a Mongoc.BSONReader pointing to the file and will parse file contents to BSON documents.

source

Mongoc.read_bson — Method.

read_bson(data::Vector{UInt8}) :: Vector{BSON}

Parses a vector of bytes to a vector of BSON documents. Useful when reading BSON as binary from a stream.

source

Mongoc.read_bson — Method.

read_bson(io::IO) :: Vector{BSON}

Reads all BSON documents from io. This method will continue to read from io until it reaches eof.

source

Mongoc.read_bson — Method.

read_bson(reader::BSONReader) :: Vector{BSON}

Reads all BSON documents from a reader.

source

Mongoc.read_next_bson — Method.

read_next_bson(reader::BSONReader) :: Union{Nothing, BSON}

Reads the next BSON document available in the stream pointed by reader. Returns nothing if reached the end of the stream.

source

Mongoc.remove_user — Method.

remove_user(database::Database, username::String)

Removes a user from database.

source

Mongoc.set_appname! — Method.

set_appname!(client::Client, appname::String)

Sets the application name for this client.

This string, along with other internal driver details, is sent to the server as part of the initial connection handshake.

C API

mongoc_client_set_appname.

source

Mongoc.transaction — Method.

transaction(f::Function, client::Client; session_options::SessionOptions=SessionOptions())

Use do-syntax to execute a transaction.

Transaction will be commited automatically. If an error occurs, the transaction is aborted.

The session parameter should be treated the same way as a Client: from a session you get a database, and a collection that are bound to the session.

Mongoc.transaction(client) do session
    database = session["my_database"]
    collection = database["my_collection"]
    new_item = Mongoc.BSON()
    new_item["inserted"] = true
    push!(collection, new_item)
end

source

Mongoc.write_bson — Method.

write_bson(io::IO, bson_list::Vector{BSON}; initial_buffer_capacity::Integer=DEFAULT_BSON_WRITER_BUFFER_CAPACITY)

Writes a vector of BSON documents to io in binary format.

Example

list = Vector{Mongoc.BSON}()

let
    src = Mongoc.BSON()
    src["id"] = 1
    src["name"] = "1st"
    push!(list, src)
end

let
    src = Mongoc.BSON()
    src["id"] = 2
    src["name"] = "2nd"
    push!(list, src)
end

open("documents.bson", "w") do io
    Mongoc.write_bson(io, list)
end

source

Mongoc.write_bson — Method.

write_bson(io::IO, bson::BSON; initial_buffer_capacity::Integer=DEFAULT_BSON_WRITER_BUFFER_CAPACITY)

Writes a single BSON document to io in binary format.

source